Restructure e2e project: analytics and onboarding
