+++ plugins/woocommerce/tests/e2e-pw/tests/analytics/analytics-access.spec.js
@@ -5,7 +5,6 @@ const { ADMIN_STATE_PATH } = require( '../../playwright.config' );
 test.describe( 'WooCommerce Home', () => {
 	test.use( { storageState: ADMIN_STATE_PATH } );
 
-	// TODO (E2E Audit): This test should be combined with other WC Homepage setup tests like the tests in activate-and-setup/task-list.spec.js into a single spec.
 	test(
 		'Can access Analytics Reports from Stats Overview',
 		{ tag: [ tags.NOT_E2E ] },
