+++ plugins/woocommerce/tests/e2e-pw/tests/analytics/analytics-data.spec.js
