+++ plugins/woocommerce/tests/e2e-pw/tests/analytics/analytics-overview.spec.js
