+++ plugins/woocommerce/tests/e2e-pw/tests/onboarding/launch-your-store.spec.js
@@ -1,8 +1,8 @@
 const { request } = require( '@playwright/test' );
 const { test, expect, tags } = require( '../../fixtures/fixtures' );
 const { setOption } = require( '../../utils/options' );
+const { activateTheme, DEFAULT_THEME } = require( '../../utils/themes' );
 const { ADMIN_STATE_PATH } = require( '../../playwright.config' );
-const wcApi = require( '@woocommerce/woocommerce-rest-api' ).default;
 
 test.describe(
 	'Launch Your Store - logged in',
@@ -10,13 +10,8 @@ test.describe(
 	() => {
 		test.use( { storageState: ADMIN_STATE_PATH } );
 
-		test.beforeEach( async ( { baseURL } ) => {
-			await new wcApi( {
-				url: baseURL,
-				consumerKey: process.env.CONSUMER_KEY,
-				consumerSecret: process.env.CONSUMER_SECRET,
-				version: 'wc-admin',
-			} ).post( 'onboarding/profile', {
+		test.beforeEach( async ( { api } ) => {
+			await api.post( 'onboarding/profile', {
 				skipped: true,
 			} );
 		} );
@@ -283,3 +278,114 @@ test.describe(
 		} );
 	}
 );
+
+async function runComingSoonTests( themeContext = '' ) {
+	const testSuffix = themeContext ? ` (${ themeContext })` : '';
+
+	test( `Entire site coming soon mode${ testSuffix }`, async ( {
+		page,
+		baseURL,
+	} ) => {
+		try {
+			await setOption(
+				request,
+				baseURL,
+				'woocommerce_coming_soon',
+				'yes'
+			);
+			await setOption(
+				request,
+				baseURL,
+				'woocommerce_store_pages_only',
+				'no'
+			);
+		} catch ( error ) {
+			console.log( error );
+		}
+
+		await page.goto( './' );
+
+		await page
+			.locator( '.woocommerce-coming-soon-banner' )
+			.waitFor( { state: 'visible' } );
+
+		await expect(
+			page.getByText(
+				"Pardon our dust! We're working on something amazing — check back soon!"
+			)
+		).toBeVisible();
+	} );
+
+	test( `Store only coming soon mode${ testSuffix }`, async ( {
+		page,
+		baseURL,
+	} ) => {
+		try {
+			await setOption(
+				request,
+				baseURL,
+				'woocommerce_coming_soon',
+				'yes'
+			);
+			await setOption(
+				request,
+				baseURL,
+				'woocommerce_store_pages_only',
+				'yes'
+			);
+		} catch ( error ) {
+			console.log( error );
+		}
+		await page.goto( 'shop/' );
+
+		await expect(
+			page.getByText( 'Great things are on the horizon' )
+		).toBeVisible();
+		await expect(
+			page.getByText(
+				'Something big is brewing! Our store is in the works and will be launching soon!'
+			)
+		).toBeVisible();
+	} );
+}
+
+test.describe( 'Launch Your Store front end - logged out', () => {
+	test.afterAll( async ( { baseURL } ) => {
+		try {
+			await setOption(
+				request,
+				baseURL,
+				'woocommerce_coming_soon',
+				'no'
+			);
+		} catch ( error ) {
+			console.log( error );
+		}
+	} );
+
+	test.describe( 'Block Theme (Twenty Twenty Four)', () => {
+		test.beforeAll( async ( { baseURL } ) => {
+			await activateTheme( baseURL, 'twentytwentyfour' );
+		} );
+
+		test.afterAll( async ( { baseURL } ) => {
+			// Reset theme to the default.
+			await activateTheme( baseURL, DEFAULT_THEME );
+		} );
+
+		runComingSoonTests( test.step, test.use );
+	} );
+
+	test.describe( 'Classic Theme (Storefront)', () => {
+		test.beforeAll( async ( { baseURL } ) => {
+			await activateTheme( baseURL, 'storefront' );
+		} );
+
+		test.afterAll( async ( { baseURL } ) => {
+			// Reset theme to the default.
+			await activateTheme( baseURL, DEFAULT_THEME );
+		} );
+
+		runComingSoonTests( test.step, test.use, 'Classic Theme (Storefront)' );
+	} );
+} );
