+++ plugins/woocommerce/tests/e2e-pw/tests/onboarding/onboarding-wizard.spec.js
