+++ plugins/woocommerce/tests/e2e-pw/tests/onboarding/setup-checklist.spec.js
